Adding in slow camera support based on old MEC python.
Creates an Ophyd device that checks the DAQ partition for detectors classified
as slow cameras and attempts to configure the devices that are found.

The detectors are not configured via the command line. They simply need to be
staged prior to running a scan, and fed the current laser configuration.

This likely needs some work, but this is the support we created while preparing
for an experiment. I'm sure it could do with some re-factoring.